% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyslog.R
\name{open_syslog}
\alias{open_syslog}
\alias{syslog}
\alias{close_syslog}
\title{Write Messages to the System Log}
\usage{
open_syslog(name, open_immediately = FALSE, include_pid = FALSE,
  errors_to_console = FALSE)

syslog(message, level = "INFO")

close_syslog()
}
\arguments{
\item{name}{A string identifying the application.}

\item{open_immediately}{When \code{TRUE}, the connection will be opened
immediately. Otherwise it will be opened when the first message is
written to the log.}

\item{include_pid}{When \code{TRUE}, include the process ID in the log
message. This may be redundant.}

\item{errors_to_console}{Write to the console if there is an error while
sending to the system logger.}

\item{message}{The message to write to the system log.}

\item{level}{The priority level of the message. One of \code{"DEBUG"},
\code{"INFO"}, \code{"NOTICE"}, \code{"WARNING"}, \code{"ERR"},
\code{"CRITICAL"}, \code{"ALERT"}, or \code{"EMERGE"} -- in that order of
priority. See \href{RFC 5424: https://tools.ietf.org/html/rfc5424#page-11}{RFC 5424}
for the basis of this schema.}
}
\description{
Write messages to the system log via the POSIX syslog interface.
}
\examples{
\dontrun{
open_syslog("my_script")
syslog("Running script.", level = "INFO")
syslog("Possible issue.", level = "WARNING")
close_syslog()

# Opening the syslog is not strictly necessary. You can
# simply write a message and it will open the log with the
# process name (likely "R") as the default.

syslog("Hello from R!", level = "WARNING")
close_syslog()
}

}
