% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyslog.R
\name{open_syslog}
\alias{open_syslog}
\alias{syslog}
\alias{close_syslog}
\title{Write Messages to the System Log}
\usage{
open_syslog(name, open_immediately = FALSE, include_pid = FALSE,
  errors_to_console = FALSE)

syslog(message, level = "INFO")

close_syslog()
}
\description{
Write Messages to the System Log
}
